+++ tests/Header/CookieParameter/ExpiresTest.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Tests\Innmind\Http\Header\CookieParameter;
 
-use Innmind\Http\Header\{
-    CookieParameter\Expires,
-    Parameter
-};
+use Innmind\Http\Header\SetCookie\Expires;
 use Innmind\TimeContinuum\PointInTime;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -14,11 +11,10 @@ class ExpiresTest extends TestCase
 {
     public function testInterface()
     {
-        $expires = new Expires(PointInTime::at(
+        $expires = Expires::at(PointInTime::at(
             new \DateTimeImmutable('2018-01-01T12:13:14+0200'),
         ));
 
-        $this->assertInstanceOf(Parameter::class, $expires);
-        $this->assertSame('Expires="Mon, 01 Jan 2018 10:13:14 GMT"', $expires->toString());
+        $this->assertSame('Expires="Mon, 01 Jan 2018 10:13:14 GMT"', $expires->toParameter()->toString());
     }
 }