+++ tests/Header/CookieParameter/PathTest.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Tests\Innmind\Http\Header\CookieParameter;
 
-use Innmind\Http\Header\{
-    CookieParameter\Path,
-    Parameter
-};
+use Innmind\Http\Header\SetCookie\Path;
 use Innmind\Url\Path as UrlPath;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -14,9 +11,8 @@ class PathTest extends TestCase
 {
     public function testInterface()
     {
-        $path = new Path(UrlPath::of('/foo'));
+        $path = Path::of(UrlPath::of('/foo'));
 
-        $this->assertInstanceOf(Parameter::class, $path);
-        $this->assertSame('Path=/foo', $path->toString());
+        $this->assertSame('Path=/foo', $path->toParameter()->toString());
     }
 }