+++ tests/Header/CookieParameter/SameSiteTest.php
@@ -3,24 +3,19 @@ declare(strict_types = 1);
 
 namespace Tests\Innmind\Http\Header\CookieParameter;
 
-use Innmind\Http\Header\{
-    CookieParameter\SameSite,
-    Parameter
-};
+use Innmind\Http\Header\SetCookie\Directive;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class SameSiteTest extends TestCase
 {
     public function testInterface()
     {
-        $sameSite = SameSite::strict();
+        $sameSite = Directive::strictSameSite;
 
-        $this->assertInstanceOf(Parameter::class, $sameSite);
-        $this->assertSame('SameSite=Strict', $sameSite->toString());
+        $this->assertSame('SameSite=Strict', $sameSite->toParameter()->toString());
 
-        $sameSite = SameSite::lax();
+        $sameSite = Directive::laxSameSite;
 
-        $this->assertInstanceOf(Parameter::class, $sameSite);
-        $this->assertSame('SameSite=Lax', $sameSite->toString());
+        $this->assertSame('SameSite=Lax', $sameSite->toParameter()->toString());
     }
 }